requirements.txt
Django<1.7
South
djangorestframework
django-ordered-model
gunicorn
raven
nose
django-nose
coverage
psycopg2
celery
redis
django-debug-toolbar
django-celery
django-celery-email
dj-database-url
django-snappy-vumi-bouncer
flake8
